speedtest-app/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
DateTime==4.3
idna==2.9
influxdb==5.3.0
msgpack==0.6.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
schedule==0.6.0
six==1.14.0
speedtest-cli==2.1.2
urllib3==1.25.9
zope.interface==5.1.0
